Re_3DVG-Small/requirements.txt
numpy==1.22.0
open3d==0.17.0
torch==1.13.1+cu117
typing==3.7.4.3
transformers==4.30.1
pillow==9.5.0
ultralytics==8.0.124
opencv-python==4.3.0.38
tqdm==4.65.0
fvcore==0.1.5.post20221221
pycocotools==2.0.6
matplotlib==3.7.1
cog==0.8.0
torchvision==0.14.1+cu117
timm==0.9.2
scipy==1.10.1
multiscaledeformableattention==1.0
setuptools==67.8.0
tabulate==0.9.0
shapely==2.0.1
datasets==2.13.1
mmengine==0.8.1
mmdet==2.25.2
imageio==2.31.1
onnx==1.14.0
requests==2.31.0
seaborn==0.12.2
cityscapesscripts==2.2.2
six==1.16.0
terminaltables==3.1.10
scikit-learn==1.3.0
pyyaml==6.0
pandas==1.3.5
rich==13.4.2